refactor: 전반적인 패키지 구조 변경
