+++ app/src/main/java/com/techcourse/ManualHandlerMapping.java
@@ -1,8 +1,8 @@
 package com.techcourse;
 
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.mvc.asis.ForwardController;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
+import com.interface21.webmvc.servlet.handler.mapping.HandlerMapping;
+import com.interface21.webmvc.servlet.mvc.Controller;
+import com.interface21.webmvc.servlet.mvc.ForwardController;
 import com.techcourse.controller.LoginController;
 import com.techcourse.controller.LoginViewController;
 import com.techcourse.controller.LogoutController;
