+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/ForwardController.java
@@ -1,6 +1,6 @@
-package com.interface21.webmvc.servlet.mvc.asis;
+package com.interface21.webmvc.servlet.mvc;
 
-import com.interface21.webmvc.servlet.ModelAndView;
+import com.interface21.webmvc.servlet.view.ModelAndView;
 import com.interface21.webmvc.servlet.view.JspView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
